Fix build failure
TEA.xs: In function ‘XS_Crypt__TEA_crypt’:
TEA.xs:58:9: error: invalid use of void expression
         if (SvREADONLY(output) || !SvUPGRADE(output, SVt_PV))
         ^~

As explained in perl5180delta.pod, this is because SvUPGRADE has changed
from an expression to a statement with no return value. In case of any
error, it will croak, so we don't have to.